+++ docs/.vitepress/theme/index.js
@@ -1,13 +1,15 @@
 // .vitepress/theme/index.js
 import DefaultTheme from 'vitepress/theme';
 import './theme.css';
-import LivePreview from './components/LivePreview.vue';
 import VueGtag from "vue-gtag";
+import LivePreview from './components/LivePreview.vue';
+import StackblitzEmbed from './components/StackblitzEmbed.vue';
 
 export default {
   ...DefaultTheme,
-  enhanceApp({app, router}) {
+  enhanceApp({ app, router }) {
     app.component('LivePreview', LivePreview);
+    app.component('StackblitzEmbed', StackblitzEmbed);
 
     // GA
     app.use(VueGtag, {