+++ packages/editor/src/vue.d.ts
@@ -4,4 +4,6 @@ declare module '*.vue' {
   export default component;
 }
 
+declare module '@morpheme/design-tokens/*';
+
 declare module '@ckeditor/ckeditor5-vue';