+++ packages/forms/src/input/SelectOptions.vue
@@ -2,6 +2,8 @@
 export interface Option {
   label: string;
   value: string | number;
+  disabled?: boolean;
+  selected?: boolean;
   options?: Option[];
 }
 
@@ -20,12 +22,20 @@ withDefaults(defineProps<Props>(), {
       <option
         v-for="(subOption, index) in option.options"
         :key="index"
-        :value="subOption?.value || subOption"
+        :value="subOption?.value"
+        :disabled="subOption?.disabled"
+        :selected="subOption?.selected"
       >
         {{ subOption?.label || subOption }}
       </option>
     </optgroup>
-    <option v-else :key="index" :value="option?.value || option">
+    <option
+      v-else
+      :key="index"
+      :value="option?.value"
+      :disabled="option?.disabled"
+      :selected="option?.selected"
+    >
       {{ option?.label || option }}
     </option>
   </template>