+++ packages/icon/src/index.ts
@@ -1,5 +1,6 @@
 import Icon from './Icon.vue';
 import './Icon.dark.scss';
 
+export * from './storage';
 export {Icon};
 export default Icon;