+++ packages/layouts/src/Grid.stories.ts
@@ -1,20 +1,20 @@
-import { VSelect } from '@morpheme/select';
-import { alignItems, alignContent, justifyContent } from './grid';
-import { VCard } from '@morpheme/card';
-import { Story } from '@storybook/vue3';
+import {VSelect} from '@morpheme/select';
+import {alignItems, flexWrap, justifyContent} from './grid';
+import {VCard} from '@morpheme/card';
+import {StoryFn} from '@storybook/vue3';
 import VRow from './VRow.vue';
 import VCol from './VCol.vue';
-import { reactive } from 'vue';
+import {reactive, ref} from 'vue';
 
 export default {
   title: 'Components/Grid',
   component: VRow,
 };
 
-export const Default: Story<{}> = (args) => ({
-  components: { VRow, VCol },
+export const Default: StoryFn = (args) => ({
+  components: {VRow, VCol},
   setup() {
-    return { args };
+    return {args};
   },
   template: `
   <VRow gap="4">
@@ -25,98 +25,126 @@ export const Default: Story<{}> = (args) => ({
   `,
 });
 
-export const Cols: Story<{}> = (args) => ({
-  components: { VRow, VCol },
+export const Cols: StoryFn = (args) => ({
+  components: {VRow, VCol},
   setup() {
-    return { args };
+    return {args};
   },
   template: `
   <VRow gap="4">
-    <VCol cols="1" class="bg-gray-100 text-center">
+    <VCol cols="1" class="bg-red-300 text-center">
       1
     </VCol>
-    <VCol cols="3" class="bg-gray-100 text-center">
+    <VCol cols="3" class="bg-green-300 text-center">
       3
     </VCol>
-    <VCol cols="8" class="bg-gray-100 text-center">
+    <VCol cols="8" class="bg-blue-300 text-center">
       8
     </VCol>
   </VRow>
   `,
 });
 
-export const Offset: Story<{}> = (args) => ({
-  components: { VRow, VCol },
+export const Offset: StoryFn = (args) => ({
+  components: {VRow, VCol},
   setup() {
-    return { args };
+    return {args};
   },
   template: `
   <VRow gap="4">
-    <VCol cols="3" class="bg-gray-100 text-center">
+    <VCol cols="3" class="bg-red-100 text-center">
       1
     </VCol>
-    <VCol cols="3" class="bg-gray-100 text-center">
+    <VCol cols="3" class="bg-green-100 text-center">
       2
     </VCol>
-    <VCol cols="3" offset="3" class="bg-gray-100 text-center">
+    <VCol cols="3" offset="3" class="bg-blue-100 text-center">
       3
     </VCol>
   </VRow>
   `,
 });
 
-export const Order: Story<{}> = (args) => ({
-  components: { VRow, VCol },
+export const Order: StoryFn = (args) => ({
+  components: {VRow, VCol},
   setup() {
-    return { args };
+    return {args};
   },
   template: `
   <VRow gap="4">
-    <VCol cols="3" order="2" class="bg-gray-100 text-center">
+    <VCol cols="4" order="2" class="bg-red-100 text-center">
       1
     </VCol>
-    <VCol cols="3" order="3" class="bg-gray-100 text-center">
+    <VCol cols="4" order="3" class="bg-green-100 text-center">
       2
     </VCol>
-    <VCol cols="3" order="1" class="bg-gray-100 text-center">
+    <VCol cols="4" order="1" class="bg-blue-100 text-center">
       3
     </VCol>
   </VRow>
   `,
 });
 
-export const Alignment: Story<{}> = (args) => ({
-  components: { VRow, VCol, VSelect },
+function toItems(obj: any) {
+  return Object.keys(obj).map((key) => ({
+    text: obj[key],
+    value: key,
+  }));
+}
+
+export const Wrap: StoryFn = (args) => ({
+  components: {VRow, VCol, VSelect},
+  setup() {
+    const wrap = ref('');
+    return {
+      args,
+      wrap,
+      wrapItems: toItems(flexWrap),
+    };
+  },
+  template: `
+  <VSelect wrapper-class="mb-4" v-model="wrap" label="Wrap" :items="wrapItems" />
+  
+  <VRow gap="4" :wrap="wrap">
+    <VCol cols="4" order="2" class="bg-red-100 text-center">
+      1
+    </VCol>
+    <VCol cols="4" order="3" class="bg-green-100 text-center">
+      2
+    </VCol>
+    <VCol cols="4" order="1" class="bg-blue-100 text-center">
+      3
+    </VCol>
+  </VRow>
+  `,
+});
+
+export const Alignment: StoryFn = (args) => ({
+  components: {VRow, VCol, VSelect},
   setup() {
     const option = reactive({
       align: 'center',
       alignContent: 'center',
       justifyContent: 'center',
-    })
-    function toItems(obj: any) {
-      return Object.keys(obj).map(key => ({
-        text: obj[key],
-        value: key
-      }));
-    }
+    });
 
     return {
       args,
       alignItems: toItems(alignItems),
       justifyContent: toItems(justifyContent),
       alignContent: toItems(justifyContent),
-      option
+      option,
     };
   },
   template: `
   <VRow gap="4" class="mb-6">
-    <VCol cols="3">
+    <VCol>
       <VSelect v-model="option.align" label="Align Items" :items="alignItems" />
     </VCol>
-    <VCol cols="3">
+    <VCol>
       <VSelect v-model="option.alignContent" label="Align Content" :items="alignContent" />
     </VCol>
-    <VCol cols="3">
+    <VCol>
       <VSelect v-model="option.justifyContent" label="Justify Content" :items="justifyContent" />
     </VCol>
   </VRow>
@@ -125,25 +153,25 @@ export const Alignment: Story<{}> = (args) => ({
     gap="4"
     v-bind="option"
   >
-    <VCol cols="3" class="bg-gray-100 text-center">
-      <div class="h-10 bg-red-500">
+    <VCol class="bg-red-100 text-center">
+      <div class="h-10 bg-purple-500">
         1
       </div>
     </VCol>
-    <VCol cols="3" class="bg-gray-100 text-center">
+    <VCol class="bg-green-100 text-center">
       2
     </VCol>
-    <VCol cols="3" class="bg-gray-100 text-center">
+    <VCol class="bg-blue-100 text-center">
       3
     </VCol>
   </VRow>
   `,
 });
 
-export const Responsive: Story<{}> = (args) => ({
-  components: { VRow, VCol, VCard },
+export const Responsive: StoryFn = (args) => ({
+  components: {VRow, VCol, VCard},
   setup() {
-    return { args };
+    return {args};
   },
   template: `
   <VRow>
@@ -161,6 +189,48 @@ export const Responsive: Story<{}> = (args) => ({
       <div class="bg-gray-200 p-2 m-2">
         3
       </div>
+    </VCol>
+  </VRow>
+  `,
+});
+
+export const LayoutExample: StoryFn = (args) => ({
+  components: {VRow, VCol, VCard},
+  setup() {
+    const stats = [
+      {
+        title: 'Total Subscribers',
+        value: '71,897',
+        color: '',
+      },
+      {
+        title: 'Avg. Open Rate',
+        value: '58.16%',
+      },
+      {
+        title: 'Avg. Click Rate',
+        value: '24.57%',
+      },
+      {
+        title: 'Avg. Bounce Rate',
+        value: '4.24%',
+      },
+    ];
+    return {args, stats};
+  },
+  template: `
+  <VRow gap="4">
+    <VCol
+      cols="12"
+      sm="3"
+      v-for="stat in stats"
+      :key="stat.title"
+    >
+      <VCard :title="stat.title">
+        <div class="text-xl font-bold">
+          {{ stat.value }}
+        </div>
+      </VCard>
     </VCol>
   </VRow>
   `,