+++ packages/select/src/vue.d.ts
@@ -6,3 +6,5 @@ declare module '*.vue' {
 
 // @headlessui-float/vue
 declare module '@headlessui/vue';
+declare module '@morpheme/design-tokens/*';
+declare module '@morpheme/forms';