+++ packages/steppers/src/Stepper.stories.ts
@@ -1,6 +1,5 @@
 import {Args, Story} from '@storybook/vue3';
 import VStepper from './Stepper.vue';
-import vueRouter from 'storybook-vue3-router';
 import {defineComponent, ref, computed, toRefs, PropType} from 'vue';
 import {useRoute} from 'vue-router';
 
@@ -26,29 +25,31 @@ const Stepper = defineComponent({
   props: {
     id: {
       type: String,
-      required: true
+      required: true,
     },
     disableRouteActive: {
       type: Boolean,
-      required: true
+      required: true,
     },
     value: {
       type: Number,
-      required: true
+      required: true,
     },
     items: {
       type: Array as PropType<Record<any, string>[]>,
-      required: true
-    }
+      required: true,
+    },
   },
-  setup(props:any) {
+  setup(props: any) {
     const {items} = toRefs(props);
     const route = useRoute();
 
     const getCurrentIdx = computed(() => {
-      return items.value?.findIndex((e:Record<string, any>) => e?.path === route.path);
+      return items.value?.findIndex(
+        (e: Record<string, any>) => e?.path === route.path,
+      );
     });
-    const getNext = computed(() =>{
+    const getNext = computed(() => {
       const currIndex = getCurrentIdx.value;
       const targetIdx = currIndex + 1;
       return targetIdx < items.value?.length ? items.value?.at(targetIdx) : {};
@@ -78,7 +79,7 @@ const Stepper = defineComponent({
         </template>
       </div>
     </div>
-  `
+  `,
 });
 
 const defaultRoutes = [
@@ -87,8 +88,8 @@ const defaultRoutes = [
     name: 'stepper',
     props: true,
     component: Stepper,
-  }
-]
+  },
+];
 
 const Template: Story = (args, ctx) => ({
   components: {
@@ -99,15 +100,15 @@ const Template: Story = (args, ctx) => ({
     const disableRouteActive = ref(args.disableRouteActive);
 
     const onPrevClick = () => {
-      if(disableRouteActive.value){
+      if (disableRouteActive.value) {
         val.value -= 1;
       }
-    }
+    };
     const onNextClick = () => {
-      if(disableRouteActive.value){
+      if (disableRouteActive.value) {
         val.value += 1;
       }
-    }
+    };
 
     return {args, onPrevClick, onNextClick, val, disableRouteActive};
   },
@@ -133,7 +134,6 @@ const Template: Story = (args, ctx) => ({
   `,
 });
 
-
 export const Default = Template.bind({});
 Default.args = {};
 Default.parameters = {
@@ -143,24 +143,15 @@ Default.parameters = {
     },
   },
 };
-Default.decorators = [
-    (story, ctx) => {
-        const initialRoute = ctx.args?.items?.at(ctx.args.modelValue)?.path
-        return vueRouter(defaultRoutes, {
-            initialRoute
-        })(story, ctx)
-    },
-];
 
-
-export const Linear = (args:Args) => ({
+export const Linear = (args: Args) => ({
   components: {
     VStepper,
   },
   setup() {
     const val = ref(args?.modelValue || 0);
 
-    const linear = [true, false]
+    const linear = [true, false];
 
     const nuArgs = {
       args: linear.map((e) => {
@@ -168,20 +159,20 @@ export const Linear = (args:Args) => ({
           ...args,
           modelValue: val,
           disableRouteActive: true,
-          linear : e,
-        }
+          linear: e,
+        };
       }),
-      linear
-    }
+      linear,
+    };
 
     const onPrevClick = () => {
-        val.value -= 1;
-    }
+      val.value -= 1;
+    };
     const onNextClick = () => {
       val.value += 1;
-    }
+    };
 
-    return {args:nuArgs, linear, val, onPrevClick, onNextClick};
+    return {args: nuArgs, linear, val, onPrevClick, onNextClick};
   },
   template: `
     <div class="flex flex-col gap-4">
@@ -212,7 +203,7 @@ Linear.parameters = {
   },
 };
 
-export const Clickable = (args:Args) => ({
+export const Clickable = (args: Args) => ({
   components: {
     VStepper,
   },
@@ -231,25 +222,32 @@ export const Clickable = (args:Args) => ({
           disableRouteActive: disableRouteActive[idx],
           clickable: clickable[idx],
           linkable: linkable[idx],
-        }
+        };
       }),
-      clickable
-    }
+      clickable,
+    };
 
     const onPrevClick = () => {
-        val.value -= 1;
-    }
+      val.value -= 1;
+    };
     const onNextClick = () => {
       val.value += 1;
-    }
+    };
 
-    const onStepClick = (step:any) => {
+    const onStepClick = (step: any) => {
       alert('You clicked a step! ' + JSON.stringify(step));
 
       val.value = step.index;
-    }
+    };
 
-    return {args:nuArgs, clickable, val, onPrevClick, onNextClick, onStepClick};
+    return {
+      args: nuArgs,
+      clickable,
+      val,
+      onPrevClick,
+      onNextClick,
+      onStepClick,
+    };
   },
   template: `
     <div class="flex flex-col gap-4">
@@ -286,7 +284,6 @@ Clickable.parameters = {
   },
 };
 
-
 export const DisableRouteActive = Template.bind({});
 DisableRouteActive.args = {
   disableRouteActive: true,
@@ -310,15 +307,6 @@ Linkable.parameters = {
     },
   },
 };
-Linkable.decorators = [
-    (story, ctx) => {
-        const initialRoute = ctx.args?.items?.at(ctx.args.modelValue)?.path
-        return vueRouter(defaultRoutes, {
-            initialRoute
-        })(story, ctx)
-    },
-];
-
 
 export const Vertical = Template.bind({});
 Vertical.args = {
@@ -331,7 +319,6 @@ Vertical.parameters = {
     },
   },
 };
-
 
 export const VModel = Template.bind({});
 VModel.args = {
@@ -347,7 +334,7 @@ VModel.args = {
       subtitle: 'Subtitle 2',
       path: '/step/2',
     },
-  ]
+  ],
 };
 VModel.parameters = {
   docs: {
@@ -356,11 +343,3 @@ VModel.parameters = {
     },
   },
 };
-VModel.decorators = [
-  (story, ctx) => {
-    const initialRoute = ctx.args?.items?.at(ctx.args.modelValue)?.path
-    return vueRouter(defaultRoutes, {
-      initialRoute
-    })(story, ctx)
-  },
-];